requirements.txt
streamlit
faiss-cpu
numpy
pandas
pyyaml
beautifulsoup4 
tqdm
requests
google-generativeai